Time: 0 ms (100%), Space: 14.7 MB (35.03%) - LeetHub

+++ 0387-first-unique-character-in-a-string/0387-first-unique-character-in-a-string.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int firstUniqChar(string s) {
         int ans = -1;
-        vector<short> fr(26);
+        vector<int> fr(26);
         for (int i = 0; i < s.size(); i++)
             fr[s[i] - 'a']++;
         for (int i = 0; i < s.size(); i++) {
